+++ pkgs/ndarray/example/copy_operations_example.dart
@@ -19,7 +19,7 @@ void runContiguousCopyExample() {
 
   // Let\'s verify memory decoupling (modifying copy does not affect parent!)
   print('\nModifying index 0 of duplicate to 99.0...');
-  duplicate.data[0] = 99.0;
+  duplicate.data[0] = Float64(99.0);
 
   print('Parent array index 0: ${parent.data[0]}');
   print('Duplicate array index 0: ${duplicate.data[0]}');
@@ -43,7 +43,7 @@ void runStridedViewCopyExample() {
   print('Copied duplicate isContiguous: ${duplicate.isContiguous}');
 
   // Decoupled memory verification
-  duplicate.data[0] = 99.0;
+  duplicate.data[0] = Float64(99.0);
   print('\nModifying copy data[0] to 99.0...');
   print('Original parent data[0] (still 1.0): ${parent.data[0]}');
   print('🏆 Strided coordinates recursively deep copied successfully!');
