+++ pkgs/ndarray/example/indexing_example.dart
@@ -18,7 +18,7 @@ void runStaticTypedAddressingExample() {
   final cellVal = mat.getCell([0, 1]);
   print('getCell([0, 1]) -> expected 20.0: $cellVal');
 
-  mat.setCell([1, 0], 99.0);
+  mat.setCell([1, 0], Float64(99.0));
   print(
     'After setCell([1, 0], 99.0) data -> expected [10, 20, 99, 40]: ${mat.toList()}',
   );
@@ -30,7 +30,7 @@ void runStaticTypedAddressingExample() {
   final negativeMask = arr < 0.0; // returns an NDArray<int> mask
 
   // Clips all negative elements to 0.0 explicitly!
-  arr.setByMaskScalar(negativeMask, 0.0);
+  arr.setByMaskScalar(negativeMask, Float64(0.0));
   print(
     'After setByMask clipping negatives -> expected [0, 5, 0, 8]: ${arr.toList()}',
   );
@@ -41,7 +41,7 @@ void runStaticTypedAddressingExample() {
   ], DType.int32);
   final targetIndices = NDArray.fromList([0, 4], [2], DType.int32);
 
-  bigVec.setIndicesScalar(targetIndices, 999);
+  bigVec.setIndicesScalar(targetIndices, Int32(999));
   print(
     'After setIndicesScalar positions [0, 4] -> expected [999, 20, 30, 40, 999]: ${bigVec.toList()}\n',
   );
